==> 1_Download_Dataset.ipynb <==
from tqdm import tqdm
import pandas as pd
import numpy as np
import librosa 
import librosa.display
import os

# 1-Prepare Data

## Make Folder Dataset

if os.path.exists('dataset')==False:
    os.mkdir('dataset')

## Download Data

cd dataset/

### Download Coswara Dataset

!git clone https://github.com/iiscleap/Coswara-Data.git

### Extracting Coswara Dataset

rt_dir = 'Coswara-Data/'
import glob
for each in os.listdir(rt_dir):
    if os.path.isdir(os.path.join(rt_dir,each)) and each != '.git':
        print(each)
        !cat {os.path.join(rt_dir, each,'')}*.tar.gz.* > {os.path.join(rt_dir, each,'')}combined_file.tar.gz
        !tar -xzf {os.path.join(rt_dir, each,'')}combined_file.tar.gz -C {rt_dir}
subset = glob.glob('Coswara-Data/*/*.tar.gz.*') + glob.glob('Coswara-Data/*/combined_data.tar.gz')
[os.remove(x) for x in subset];

### Download Coughvid Dataset

!wget https://zenodo.org/record/4048312/files/public_dataset.zip

### Extracting Coughvid Dataset

if os.path.exists('Coughvid-Data')==False:
    os.mkdir('Coughvid-Data')

!unzip "public_dataset.zip" -d "Coughvid-Data/"
