==> 4_Training.ipynb <==
## Downloading Pre-Trained Model

!wget https://zenodo.org/record/3987831/files/MobileNetV1_mAP%3D0.389.pth

## Training 5 Seeds

import os

config = 'script/config.json'
seeds = [42]
for seed in seeds:
    command = f"python script/train.py -c {config} -s {seed}"
    os.system(command)
