==> 5_Testing.ipynb <==
## Testing 5 Seed

import os

data = 'csv_files/experiment_data/test.csv'
folder = 'dataset/Merge-Data/'
experiment_dir = 'result/'
command = f"python script/test_all_seeds.py -t {data} -r {folder} --experiment_dir {experiment_dir}"
os.system(command)

import os

data = 'csv_files/experiment_data/test.csv'
folder = 'dataset/Merge-Data/'
experiment_dir = 'result/'
musan = 'dataset/Musan-Data'
command = f"python script/test_all_seeds.py -t {data} -r {folder} --experiment_dir {experiment_dir} --insert_noise 1 --num_additive 1 --num_specaug 1 --musan_path {musan}"
os.system(command)
